Rewrite ClaspyRun.__init__, with command part and setting argument
